requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via wmctrl
black==25.1.0
    # via -r requirements.in
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via black
cryptography==44.0.2
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==4.2.20
    # via
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   social-auth-app-django
django-cors-headers==4.7.0
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.0
    # via djoser
djoser==2.3.1
    # via -r requirements.in
factory-boy==3.3.3
    # via -r requirements.in
faker==37.1.0
    # via factory-boy
fancycompleter==0.9.1
    # via pdbpp
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
invoke==2.2.0
    # via -r requirements.in
isort==6.0.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via black
numpy==2.2.4
    # via pandas
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==24.2
    # via
    #   black
    #   pytest
pandas==2.2.3
    # via -r requirements.in
pathspec==0.12.1
    # via black
pdbpp==0.10.3
    # via -r requirements.in
pillow==11.1.0
    # via -r requirements.in
platformdirs==4.3.7
    # via black
pluggy==1.5.0
    # via pytest
pycparser==2.22
    # via cffi
pygments==2.19.1
    # via pdbpp
pyjwt==2.9.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyrepl==0.9.0
    # via fancycompleter
pytest==8.3.5
    # via
    #   pytest-django
    #   pytest-repeat
pytest-django==4.11.1
    # via -r requirements.in
pytest-repeat==0.9.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python3-openid==3.2.0
    # via social-auth-core
pytz==2025.2
    # via pandas
requests==2.32.3
    # via
    #   requests-oauthlib
    #   social-auth-core
    #   stripe
requests-oauthlib==2.0.0
    # via social-auth-core
six==1.17.0
    # via python-dateutil
social-auth-app-django==5.4.3
    # via djoser
social-auth-core==4.5.6
    # via social-auth-app-django
sqlparse==0.5.3
    # via django
stripe==11.6.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via stripe
tzdata==2025.2
    # via
    #   faker
    #   pandas
urllib3==2.3.0
    # via requests
wmctrl==0.5
    # via pdbpp
